feat(client): Refactor move charts to common folder
