+++ client/src/app/components/stardust/statistics/charts/LineChart.scss
@@ -1,5 +1,5 @@
-@import "../../../../scss/mixins";
-@import "../../../../scss/variables";
+@import "../../../../../scss/mixins";
+@import "../../../../../scss/variables";
 
 .line-chart--wrapper {
     position: relative;
