+++ client/src/app/components/stardust/statistics/charts/StackedBarChart.scss
@@ -1,18 +1,20 @@
-@import "../../../../scss/mixins";
-@import "../../../../scss/variables";
+@import "../../../../../scss/mixins";
+@import "../../../../../scss/variables";
+@import "../../../../../scss/fonts";
 
-.bar-chart--wrapper {
+.stack-bar-chart--wrapper {
     position: relative;
     width: fit-content;
     padding: 14px;
     margin-bottom: 14px;
-    background-color:  var(--light-bg);
+    background-color: $gray-daylight;
     border-radius: 8px;
 
     svg.hook {
         rect {
-            &:hover {
-                stroke: $gray-6;
+            &.hover-highlight {
+                stroke: $gray-5;
+                stroke-width: 1px;
             }
         }
