+++ client/src/app/components/stardust/statistics/charts/StackedLineChart.scss
@@ -1,6 +1,6 @@
-@import "../../../../scss/mixins";
-@import "../../../../scss/variables";
-@import "../../../../scss/fonts";
+@import "../../../../../scss/mixins";
+@import "../../../../../scss/variables";
+@import "../../../../../scss/fonts";
 
 svg.hook {
     rect {
