+++ client/src/app/components/stardust/statistics/charts/StackedLineChart.tsx
@@ -3,8 +3,8 @@ import { axisBottom, axisLeft } from "d3-axis";
 import { select } from "d3-selection";
 import moment from "moment";
 import React, { useCallback, useLayoutEffect, useRef, useState } from "react";
-import ChartHeader, { TimespanOption } from "./ChartHeader";
-import ChartTooltip from "./ChartTooltip";
+import ChartHeader, { TimespanOption } from "../ChartHeader";
+import ChartTooltip from "../ChartTooltip";
 import "./StackedLineChart.scss";
 
 interface StackedLineChartProps {
